extracted enhancement of ehrenamtsuche-hessen from default enhancement
